% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightstep.R
\name{lightstep}
\alias{lightstep}
\title{Fast and light stepwise selection of regressors}
\usage{
lightstep(data, ic = c("AICc", "AIC", "BIC", "BICc"), silent = TRUE,
  df = NULL, subset = NULL, distribution = c("dnorm", "dlaplace", "ds",
  "dgnorm", "dlogis", "dt", "dalaplace", "dlnorm", "dllaplace", "dls",
  "dlgnorm", "dbcnorm", "dinvgauss", "dgamma", "dexp", "dfnorm", "drectnorm",
  "dpois", "dnbinom", "dbeta", "dlogitnorm", "plogis", "pnorm"),
  occurrence = c("none", "plogis", "pnorm"), ...)
}
\arguments{
\item{data}{Data frame containing dependant variable in the first column and
the others in the rest. The variables need to be numeric, categorical ones are not
supported and should be expanded via the \link[stats]{model.matrix} in advance.}

\item{ic}{Information criterion to use.}

\item{silent}{If \code{silent=FALSE}, then nothing is silent, everything is
printed out. \code{silent=TRUE} means that nothing is produced.}

\item{df}{Number of degrees of freedom to add (should be used if stepwise is
used on residuals).}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{distribution}{Distribution to pass to \code{lightlm()}. See \link[lightbox]{lightlm}
for details.}

\item{occurrence}{what distribution to use for occurrence part. See
\link[lightbox]{lightlm} for details.}

\item{...}{This is temporary and is needed in order to capture "silent"
parameter if it is provided.}
}
\value{
Function returns \code{model} - the final model of the class "lightlm".
See \link[lightbox]{lightlm} for details of the output.
}
\description{
Function selects variables that give linear regression with the lowest
information criteria. The selection is done stepwise (forward) based on
partial correlations. This should be a simpler and faster implementation
than step() function from the `stats' package.
}
\details{
The algorithm uses lightlm() to fit different models and cor() to select the next
regressor in the sequence.
}
\examples{

### Simple example
obs <- 100000
xreg <- matrix(rnorm(200*obs,10,3),obs,200)
xreg <- cbind(100+0.5*xreg[,1]-0.75*xreg[,2]+rnorm(obs,0,3),xreg,rnorm(obs,300,10))
colnames(xreg) <- c("y",paste0("x",c(1:200)),"Noise")
lightstep(xreg)

}
\seealso{
\code{\link[stats]{step}, \link[greybox]{xregExpander},
\link[greybox]{stepwise}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.com}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
