% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{stepwise}
\alias{stepwise}
\title{Stepwise selection of regressors}
\usage{
stepwise(data, ic = c("AICc", "AIC", "BIC", "BICc"), silent = TRUE,
  df = NULL, formula = NULL, subset = NULL, method = c("pearson",
  "kendall", "spearman"), distribution = c("dnorm", "dlaplace", "ds",
  "dgnorm", "dlogis", "dt", "dalaplace", "dlnorm", "dllaplace", "dls",
  "dlgnorm", "dbcnorm", "dinvgauss", "dgamma", "dexp", "dfnorm", "drectnorm",
  "dpois", "dnbinom", "dbeta", "dlogitnorm", "plogis", "pnorm"),
  occurrence = c("none", "plogis", "pnorm"), ...)
}
\arguments{
\item{data}{Data frame containing dependant variable in the first column and
the others in the rest.}

\item{ic}{Information criterion to use.}

\item{silent}{If \code{silent=FALSE}, then nothing is silent, everything is
printed out. \code{silent=TRUE} means that nothing is produced.}

\item{df}{Number of degrees of freedom to add (should be used if stepwise is
used on residuals).}

\item{formula}{If provided, then the selection will be done from the listed
variables in the formula after all the necessary transformations.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{method}{Method of correlations calculation. The default is Pearson's
correlation, which should be applicable to a wide range of data in different scales.}

\item{distribution}{Distribution to pass to \code{alm()}. See \link[greybox]{alm}
for details.}

\item{occurrence}{what distribution to use for occurrence part. See
\link[greybox]{alm} for details.}

\item{...}{This is temporary and is needed in order to capture "silent"
parameter if it is provided.}
}
\value{
Function returns \code{model} - the final model of the class "alm".
See \link[greybox]{alm} for details of the output.
}
\description{
Function selects variables that give linear regression with the lowest
information criteria. The selection is done stepwise (forward) based on
partial correlations. This should be a simpler and faster implementation
than step() function from the `stats' package.
}
\details{
The algorithm uses alm() to fit different models and cor() to select the next
regressor in the sequence.
}
\examples{

### Simple example
xreg <- cbind(rnorm(100,10,3),rnorm(100,50,5))
xreg <- cbind(100+0.5*xreg[,1]-0.75*xreg[,2]+rnorm(100,0,3),xreg,rnorm(100,300,10))
colnames(xreg) <- c("y","x1","x2","Noise")
stepwise(xreg)

}
\seealso{
\code{\link[stats]{step}, \link[greybox]{xregExpander},
\link[greybox]{lmCombine}}
}
\author{
Ivan Svetunkov, \email{ivan@svetunkov.com}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{ts}
